error handling

+++ city-backend/src/controllers/cityController.js
@@ -4,14 +4,23 @@ import { City } from "../models/cities.js";
 export const getCities = async (req, res) => {
   res.set("Access-Control-Allow-Origin", "http://localhost:3000");
   const cities = await City.find();
+  if(cities.length==0){
+    return res.status(400).send({error: "Eintrag nicht gefunden."});
+  }
   res.status(200).send(cities);
 };
 export const getCityById = async (req, res) => {
   let city = await City.findById(req.params.id);
+  if(city.length==0){
+    return res.status(400).send({error: "Eintrag nicht gefunden."});
+  }
   res.status(200).send(city);
 };
 export const getCityByName = async (req, res) => {
   let result = await City.find({cityName: req.query.cityName})
+  if(result.length==0){
+    return res.status(400).send({error: "Eintrag nicht gefunden."});
+  }
   res.status(200).send(result);
 };
 
@@ -20,7 +29,12 @@ export const addCity = async (req, res) => {
   if (!errors.isEmpty()) {
     return res.status(400).json({ errors: errors.array() });
   }
-  
+
+  let result = await City.find({ cityName: req.body.cityName });
+  if(result.length!=0){
+      return res.status(400).send({error: `Eintrag schon vorhanden`});
+  }
+
   const city = new City({
     cityName: req.body.cityName,
     population: req.body.population,
@@ -34,6 +48,11 @@ export const updateCity = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(400).json({ errors: errors.array() });
+  }
+  
+  let result = await City.findById(req.params.id);
+  if(result.length==0){
+      return res.status(400).send({error: `Eintrag nicht gefunden`});
   }
 
   const city = new City({
@@ -55,6 +74,10 @@ export const updateCity = async (req, res) => {
 };
 
 export const deleteCity = async (req, res) => {
+  let result = await City.findById(req.params.id);
+  if(result.length==0){
+      return res.status(400).send({error: `Eintrag nicht gefunden`});
+  }
   await City.deleteOne(City.findById(req.params.id));
   res.status(200).send("Eintrag gelöscht")
 }
