+++ city-backend/src/controllers/countryController.js
@@ -4,16 +4,25 @@ import { Country } from "../models/countries.js";
 export const getCountries = async (req, res) => {
     res.set("Access-Control-Allow-Origin", "http://localhost:3000");
     const countries = await Country.find();
+    if(countries.length==0){
+      return res.status(400).send({error: "Eintrag nicht gefunden."});
+    }
     res.status(200).send(countries);
   };
 
   export const getCountryById = async (req, res) => {
     let country = await Country.findById(req.params.id);
+    if(country.length==0){
+      return res.status(400).send({error: "Eintrag nicht gefunden."});
+    }
     res.status(200).send(country);
   };
 
   export const getCountryByName = async (req, res) => {
     let result = await Country.find({countryName: req.query.countryName})
+    if(result.length==0){
+      return res.status(400).send({error: "Eintrag nicht gefunden."});
+    }
     res.status(200).send(result);
   };
 
@@ -21,6 +30,11 @@ export const getCountries = async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json({ errors: errors.array() });
+    }
+
+    let result = await Country.find({ countryName: req.body.countryName });
+    if(result.length!=0){
+        return res.status(400).send({error: `Eintrag schon vorhanden`});
     }
     
     const country = new Country({
@@ -37,6 +51,11 @@ export const getCountries = async (req, res) => {
         return res.status(400).json({ errors: errors.array() });
       }
       
+      let result = await Country.findById(req.params.id);
+      if(result.length==0){
+      return res.status(400).send({error: `Eintrag nicht gefunden`});
+      }
+
       const country = new Country({
           countryName: req.body.countryName,
           capital:  req.body.capital,   
@@ -56,6 +75,10 @@ export const getCountries = async (req, res) => {
       };
 
       export const deleteCountry = async (req, res) => {
+        let result = await Country.findById(req.params.id);
+        if(result.length==0){
+            return res.status(400).send({error: `Eintrag nicht gefunden`});
+        }
         await Country.deleteOne(Country.findById(req.params.id));
         res.status(200).send("Eintrag gelöscht")
       }
